add autoexplore reissue behavior
This commit is intended to workaround several issues with travel_open_doors.

When using travel_open_doors = avoid, the game silently breaks autoexplore
adjacent to closed doors. With default behavior, this requires the player to
open those doors manually. With some minor Lua ch_stop_running() tinkering,
the player can reduce that to just tapping 'o' again each time it stops.

The intent here is to try to go further, providing a near-seamless transition
between the game's 'avoid' and 'open' door behavior:

Let the player press 'o' once, and the script will handle the rest, by
beginning each explore using 'avoid', then re-issuing explores in the
background, switching between 'avoid' and 'open' as needed.

Unfortunately, there are a ton of undesirable background interactions that
make this implementation difficult.

The game doesn't expose any way for the script to know that autoexplore
has been stopped due to door avoid behavior: It just stops. As a result,
when explore stops, we just have to take our best guess about reissuing it.
This can result in infinite loops. I've placed an assert, but ugh.

We can't act from within c_message, because it can be called when it's not
the player's turn.

We can't set shared flags from c_message based on "Done exploring"; there's
a bad interaction here with ch_stop_running. I didn't fully test this and I'm
not sure if these hooks are preempting each other or if there's a race
condition or what, but it seems to be a really bad idea, generally, to use
entry points into the crawl api that aren't ready(). It's really easy to send
the game into an infinite loop and max cpu usage with these hooks.

We can act from within ch_stop_running, but we shouldn't: Most future "auto"
script actions won't be called from there, and it's a bad idea to split
up our action handling across the script.

"travel_open_doors = avoid" breaks interlevel G> behavior when near closed
doors, so we have to handle that, too.

I've done my best here for now, but I expect infinite loop issues with the
reissue logic, and I might end up partially reverting some of this later.

Looking at it as-is, I'm not convinced it's worth the code complexity, but
I really do want the single-tap 'o' behavior, so I'll try it like this.