adjust warning behavior by threat tier
Prior to this commit, the script was getting pretty spammy with force_mores.
The intent here is to reduce force_more spam for threats that don't strictly
need a force, while still issuing warnings for those lower tiers of threats
that are (hopefully) obvious enough to not be tabbed through.

This is an initial pass at rough threat tiers for each warning, I'm probably
going to need to revise some of these tiers.

Also, tier 2 might need something more than a single flash:
flash_screen_message doesn't seem to mix very well with my delay settings.

The flash shown at present is a very brief (one frame?) thing, it seems like
it's going to be easy to miss.