fix reissue autoexplore infinite loop
This one occurred when the script tried to reissue adjacent to an excluded
closed door. This might(?) be the wrong place to put the exclude check:
I can think of some future situations where I might want to check adjacent
feats without caring whether they're excluded.

For now, however, the script is only using this in one place, so I'll deal
with that in the future if it comes up.