handle generic wands
As far as I can tell, the game doesn't display success % for wands here,
but I've tried to allow the pattern match to capture that anyway in case
I'm wrong or the code is changed in future.

A side benefit of writing a dynamic reason here is that as-written, this
issues a new warning each time a monster fires a wand, which seems like good
behavior, so I'm keeping it. It's a feature, not a bug!