0.30 tournament tweak: disable some spam
The imminent death check here is inaccurate, as it doesn't check player
resistances, and it's also incredibly spammy in zot to the point of being
unplayable. I'm temporarily disabling this for now.

For this check to be properly usable, it will need a mapping of monster
ability names to beam type, and beam type to damage type, so that it could
check player resists in a generic way.

That's all pretty fragile and susceptible to bitrot, but there doesn't seem
to be a better way to do it from clua, so it's much easier to just disable
this for now.

Ideally, monster ability damage types would be exposed through lua, but it's
unclear if any other downstream users besides this script would ever make use
of that feature (qw doesn't appear to do anything like this...), so it might
be better to keep the jank here, instead of adding to crawl's lua maintenance
burden long-term.