Fix an autoexplore + autopickup loop
Prior to this commit, the script's maybe_reissue_autoexplore conditions
could send the game into a loop if autoexplore ended on a tile adjacent
to a closed door and containing an item marked for autopickup, while the
player's inventory was full.

This commit is a small fix that forces the script to stop reissuing
autoexplore in this particular case.

Our maybe_reissue_autoexplore logic is pretty bad, though, and there are
probably similar unhandled errors lurking here, where autoexplore
*should* end next to a door but instead sends itself into a loop.

Ultimately, this reissue autoexplore logic needs a rewrite at some point,
but that can happen in another commit.